requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.3
celery==5.3.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==4.2.16
django-celery-beat==2.8.1
django-extensions==3.2.3
django-timezone-field==7.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
kombu==5.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrovich==0.0.4
pytz==2024.2
redis==5.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
